requirements.txt
action-msgs==1.2.2
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.9.0
actuator-msgs==0.0.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
builtin-interfaces==1.2.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
composition-interfaces==1.2.2
cv-bridge==3.2.1
dataclasses-json==0.6.7
demo-nodes-py==0.20.5
Deprecated==1.2.18
diagnostic-msgs==4.9.0
distro==1.9.0
domain-coordinator==0.10.0
dotenv==0.9.9
duckdb==1.4.1
example-interfaces==0.9.3
examples-rclpy-executors==0.15.4
examples-rclpy-minimal-action-client==0.15.4
examples-rclpy-minimal-action-server==0.15.4
examples-rclpy-minimal-client==0.15.4
examples-rclpy-minimal-publisher==0.15.4
examples-rclpy-minimal-service==0.15.4
examples-rclpy-minimal-subscriber==0.15.4
faiss-cpu==1.12.0
frozenlist==1.8.0
geometry-msgs==4.9.0
gitdb==4.0.12
GitPython==3.1.45
gps-msgs==2.1.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
human_mouse==0.1.2
idna==3.11
image-geometry==3.2.1
interactive-markers==2.3.2
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.2.17
langchain-classic==1.0.0
langchain-community==0.2.19
langchain-core==0.2.43
langchain-experimental==0.0.42
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.1.147
laser-geometry==2.4.0
launch==1.0.10
launch-ros==0.19.10
launch-testing==1.0.10
launch-testing-ros==0.19.10
launch-xml==1.0.10
launch-yaml==1.0.10
lifecycle-msgs==1.2.2
logging-demo==0.20.5
map-msgs==2.1.0
MarkupSafe==3.0.3
marshmallow==3.26.1
message-filters==4.3.8
MouseInfo==0.1.3
mss==10.1.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.10.2
nav-msgs==4.9.0
nodriver==0.47.0
numpy==1.26.4
openai==1.109.1
opencv-python==4.12.0.88
orjson==3.11.4
ormsgpack==1.12.0
osrf-pycommon==2.1.6
packaging==24.2
pandas==2.3.3
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.0
pyarrow==21.0.0
PyAutoGUI==0.9.54
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pypdf==6.1.3
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-qt-binding==1.1.2
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.3
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.2
rclpy==3.3.17
rcutils==5.1.6
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros-gz-interfaces==0.244.20
ros2action==0.18.13
ros2bag==0.15.15
ros2cli==0.18.13
ros2component==0.18.13
ros2doctor==0.18.13
ros2interface==0.18.13
ros2launch==0.19.10
ros2lifecycle==0.18.13
ros2multicast==0.18.13
ros2node==0.18.13
ros2param==0.18.13
ros2pkg==0.18.13
ros2run==0.18.13
ros2service==0.18.13
ros2topic==0.18.13
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.5
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rpds-py==0.28.0
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scipy==1.16.2
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sros2==0.10.7
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
streamlit==1.51.0
teleop-twist-keyboard==2.4.0
tenacity==8.5.0
tf2-geometry-msgs==0.25.15
tf2-kdl==0.25.15
tf2-msgs==0.25.15
tf2-py==0.25.15
tf2-ros-py==0.25.15
tf2-sensor-msgs==0.25.15
tf2-tools==0.25.15
theora-image-transport==2.5.3
tiktoken==0.12.0
toml==0.10.2
topic-monitor==0.20.5
tornado==6.5.2
tqdm==4.67.1
trajectory-msgs==4.9.0
turtlesim==1.4.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unique-identifier-msgs==2.2.1
urllib3==2.5.0
vision-msgs==4.1.1
visualization-msgs==4.9.0
watchdog==6.0.0
websockets==15.0.1
wrapt==1.17.3
xacro==2.0.13
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
